Add '2>/dev/null' behind 'screencap -p' command
On Linux systems with Waydroid running on an AMD GPU platform, you may encounter an error message: "/vendor/etc/hwdata/amdgpu.ids: No such file or directory."  
This error is written into the header of the captured screenshot PNG file, causing the file to be considered corrupted and unreadable.  
To resolve this, append '2>/dev/null' to the command to suppress the error message, ensuring that the screenshot file remains readable for scripts.